Add : Difficulty + Time
#149
+ 이제 난이도는 시간 경과에 따라서 증가할 예정
~ 레벨 디자인이 필요.. Mapinfo 설명 추가

+++ MarstoEarth/Assets/Scripts/Infos/MapInfo.cs
@@ -1,11 +1,16 @@
 public static class MapInfo
 {
+    // 현재 맵의 시드 넘버
     public static int seed_Number = 0;
-    public static int difficulty = 0;
-    // MapManager의 초기화 이후에 1로 증가
+    // 난이도, 시간 * a + 스테이지 * a 의 비율로 증가
+    public static float difficulty = 0;
+    // cur_Stage는 MapManager의 초기화 이후에 1로 증가
     public static int cur_Stage = 0;
+    // 현재 방의 갯수
     public static int node_num = 4;
+    // 맵 생성 시 사용 될 노드 풀
     public static NodePool cur_NodePool = NodePool.All;
+    // retry 확인 용
     public static bool isRetry = false;
     // 노드 생성 시 시작 방과의 거리에 따른 방 생성 확률
     public static float maxDistance = 3f;
